A Class to configure gemrc
this class and its tests are based on ruby::config. The purpose is to
manage a global /etc/gemrc.
n.b.: We use pick for checking toggling whether or not to create a file,
however, pick doesn't needs at least one non-null value, hence the last
statement ''No need for gemrc.''